cc2538dk: Fix builds of examples with some specific paths
CONTIKI_TARGET_DIRS is relative to the target folder where the platform Makefile
is located. Using this variable for the Contiki CC2538DK platform paths broke
the build of Mist examples with deeper path depths and of out-of-tree examples.
On the contrary, PROJECTDIRS does not assume anything about the given paths, so
use it instead.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>