enc28j60: Try to prevent Ethernet -> RF overrun
The Ethernet receptions are way faster than RF transmissions, so it is very
likely that packet queue overrun conditions occur, which then results in 'drop
incoming frag' errors on mesh nodes because some Ethernet packets could not be
forwarded to the RF.

Unfortunately, there is no easy way to know if more Ethernet packets can be
queued for RF forwarding before reading them from the controller. Hence, this is
only a workaround based on timings. After reading an Ethernet packet, the delay
before reading the next one is made proportional to the packet length. In that
way, the Ethernet controller handles the flow control itself.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>