eth2538: enc28j60: Use the SPI module instead of GPIOs
According to the ENC28J60 erratum #1, when the SPI clock from the host
microcontroller is run at frequencies of less than 8 MHz, reading from or
writing to the MAC registers may be unreliable. Hence, use the CC2538 SPI module
at 8 MHz rather than slower GPIO operations to drive the SPI pins. This also
makes the code simpler, and it gives more time to the CC2538 to perform other
tasks.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>