http-socket: Use puts() instead of printf() for simple strings
puts() is more efficient than printf() for simple strings, both because the
string has not to be parsed during printing, and because the terminating new
line character has not to be stored. Hence, use puts() by default. This also
fixes the missing new line character for the "null url" and "empty url"
messages.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>