http-socket: Fix HTTP_SOCKET_OUTPUTBUFSIZE
The return values of the calls to tcp_socket_send_str() are ignored, so, in
order to keep things simple, HTTP_SOCKET_OUTPUTBUFSIZE has to be large enough to
contain the longest possible request, which is 143 characters with the current
code. Round this value up to 256 to keep room for future enhancements.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau@advansee.com>